+++ SQLBinding.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30204.135
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SQLFunction", "src\SQLFunction\SQLFunction.csproj", "{B2E0EFB9-BEDB-48F7-BD5A-0E1010333475}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SQLBindingExtension", "src\SQLExtension\SQLBindingExtension.csproj", "{4453B407-2CA3-4011-BDE5-247FCE5C1974}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{B2E0EFB9-BEDB-48F7-BD5A-0E1010333475}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B2E0EFB9-BEDB-48F7-BD5A-0E1010333475}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B2E0EFB9-BEDB-48F7-BD5A-0E1010333475}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B2E0EFB9-BEDB-48F7-BD5A-0E1010333475}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4453B407-2CA3-4011-BDE5-247FCE5C1974}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4453B407-2CA3-4011-BDE5-247FCE5C1974}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4453B407-2CA3-4011-BDE5-247FCE5C1974}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4453B407-2CA3-4011-BDE5-247FCE5C1974}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {49902AA5-150F-4567-B562-4AA8549B2CF4}
+	EndGlobalSection
+EndGlobal
