--- src/SQLExtension/SQLInputBinding.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30204.135
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SQLBindingFunction", "SQLInputBinding\SQLBindingFunction.csproj", "{ADC02480-5844-428A-913C-09D8FECA9D77}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SQLBinding", "SQLBinding\SQLBinding.csproj", "{E2BFEB7B-CB75-48DC-A984-2161FC8A407C}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{ADC02480-5844-428A-913C-09D8FECA9D77}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{ADC02480-5844-428A-913C-09D8FECA9D77}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{ADC02480-5844-428A-913C-09D8FECA9D77}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{ADC02480-5844-428A-913C-09D8FECA9D77}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E2BFEB7B-CB75-48DC-A984-2161FC8A407C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E2BFEB7B-CB75-48DC-A984-2161FC8A407C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E2BFEB7B-CB75-48DC-A984-2161FC8A407C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E2BFEB7B-CB75-48DC-A984-2161FC8A407C}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {49902AA5-150F-4567-B562-4AA8549B2CF4}
-	EndGlobalSection
-EndGlobal
