+++ src/SQLFunction/Properties/serviceDependencies.local.json
